% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_income.R
\name{get_income}
\alias{get_income}
\title{Calculate income based on time and hourly wages}
\usage{
get_income(
  hourly_income,
  interval,
  workhours_weekly = 37.5,
  workhours_monthly = 162.95,
  workhours_annual = 162.95 * 12
)
}
\arguments{
\item{hourly_income}{Hourly income in pounds}

\item{interval}{The time interval to calculate income for, which must be
"week", "month" or "annual".}

\item{workhours_weekly}{The number of work hours in a week.}

\item{workhours_monthly}{The number of work hours per month}

\item{workhours_annual}{The number of work hours per year}
}
\value{
The value of income in pounds
}
\description{
Calculate income based on time and hourly wages
}
\examples{
get_income(hourly_income = 23, interval = "week",
workhours_weekly = 28)
}
