% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workhours.R
\name{get_workhours_annual}
\alias{get_workhours_annual}
\title{Calculate the number of annual work hours for a given work pattern}
\usage{
get_workhours_annual(workhours_weekly, workweeks_annual = 365/7)
}
\arguments{
\item{workhours_weekly}{The number of hours worked each week}

\item{workweeks_annual}{The number of weeks worked each year}
}
\value{
The number of work hours in one year
}
\description{
\code{get_workhours_annual} can be used in with the output of \link{get_workhours_weekly}
as input (see Examples).
}
\examples{
get_workhours_annual(37.5)

get_workhours_annual(workhours_weekly = get_workhours_weekly(6, 4))
}
