% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workhours.R
\name{get_workhours_monthly}
\alias{get_workhours_monthly}
\title{Calculate the number of monthly work hours for a given work pattern}
\usage{
get_workhours_monthly(workhours_annual, workmonths_annual = 12)
}
\arguments{
\item{workhours_annual}{The number of hours worked each year}

\item{workmonths_annual}{The number of months worked each year}
}
\value{
The number of work hours in one month
}
\description{
Calculating using months as an interval can be confusing because months
do not have equal lengths.
}
\details{
In the NHS pay system, this complexity is removed by treating each
month as having an identical length. The number of hours worked per month
is calculated from the number of hours worked per year, and can
be combined with \code{\link[=get_workhours_annual]{get_workhours_annual()}} (see Examples).

This means to calculate intervals it is easier to think in the order:
\itemize{
\item hours
\item days
\item weeks
\item years
\item months
}
}
\examples{
get_workhours_monthly(workhours_annual =
get_workhours_annual(workhours_weekly = 37.5))
}
