% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pay_ni.R
\name{pay_ni}
\alias{pay_ni}
\title{Calculate national insurance payments}
\usage{
pay_ni(
  weekly_pay,
  standard_bracket_lower = 242,
  standard_bracket_higher = 967,
  standard_ni_rate = 0.08,
  upper_ni_rate = 0.02
)
}
\arguments{
\item{weekly_pay}{Weekly pay in pounds}

\item{standard_bracket_lower}{The lower pay threshold for the standard bracket}

\item{standard_bracket_higher}{The higher pay threshold for the standard bracket}

\item{standard_ni_rate}{The standard rate of national insurance}

\item{upper_ni_rate}{The upper rate of national insurance}
}
\value{
The amount of national insurance to pay weekly in pounds.
}
\description{
The default values for this function are taken from the
\href{https://www.gov.uk/national-insurance/how-much-you-pay}{government website}.
}
\examples{
pay_ni(350)
}
