% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pay_paye.R
\name{pay_paye}
\alias{pay_paye}
\title{Calculate income tax on Pay As You Earn (PAYE)}
\usage{
pay_paye(
  annual_pay_taxable,
  personal_allowance = 12570,
  basic_rate_upper = 50270,
  higher_rate_upper = 125140,
  initial_rate = 0.2,
  higher_rate = 0.4,
  additional_rate = 0.45
)
}
\arguments{
\item{annual_pay_taxable}{Taxable annual income in pounds}

\item{personal_allowance}{The personal allowance you don't pay income tax on}

\item{basic_rate_upper}{The upper limit for paying the basic rate of income tax}

\item{higher_rate_upper}{The upper limit for paying the higher rate of income tax}

\item{initial_rate}{The initial rate of income tax}

\item{higher_rate}{The higher rate of income tax}

\item{additional_rate}{The "additional" (i.e. higher than the higher rate) of income tax}
}
\value{
The annual amount of income tax to pay in pounds
}
\description{
This function uses default values from the PAYE \href{https://www.gov.uk/income-tax-rates}{government website}
}
\examples{
pay_paye(12570)
}
