% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pay_pension.R
\name{pay_pension}
\alias{pay_pension}
\title{Calculate the amount of pay put into your pension}
\usage{
pay_pension(monthly_pay, pension_rate = 0.098)
}
\arguments{
\item{monthly_pay}{Monthly pay amount in pounds.}

\item{pension_rate}{The pension rate, which defaults to 9.8\%.}
}
\value{
The amount to pay into your pension in pounds.
}
\description{
Calculate the amount of pay put into your pension
}
\examples{
pay_pension(3833)
}
